Extract MainSequence from within the star 💫
